+++ a2c_discrete.py
@@ -1,5 +1,4 @@
-import tr_helpers
-import wrappers
+import common.tr_helpers as tr_helpers
 import tensorflow as tf
 import numpy as np
 import collections
@@ -8,7 +7,7 @@ from collections import deque, OrderedDict
 from tensorboardX import SummaryWriter
 from tensorflow_utils import TensorFlowVariables
 import gym
-import vecenv
+import common.vecenv as vecenv
 from tf_moving_mean_std import MovingMeanStd
 from datetime import datetime
