+++ common/experience.py
@@ -1,7 +1,7 @@
 import numpy as np
 import random
 
-from segment_tree import SumSegmentTree, MinSegmentTree
+from common.segment_tree import SumSegmentTree, MinSegmentTree
 
 
 class ReplayBuffer(object):
