+++ common/segment_tree.py
