+++ common/tr_helpers.py
