+++ data.yml
@@ -6,66 +6,41 @@ config:
   critic_coef: 1
   e_clip: 0.2
   entropy_coef: 0.0
-  env_name: BipedalWalkerHardcoreCnn-v2
+  env_name: BipedalWalkerHardcore-v2
   gamma: 0.99
   grad_norm: 0.5
   learning_rate: 1e-4
   lr_schedule: adaptive
   lr_threshold: 0.008
-  max_epochs: 20000
+  max_epochs: 10000
   mini_epochs: 4
   minibatch_size: 2048
-  name: walker_0_1
+  name: walker_0_0
   normalize_advantage: true
   normalize_input: false
   num_actors: 16
   ppo: true
   reward_shaper:
-    min_val: -1
     scale_value: 0.1
-  score_to_win: 300
+  score_to_win: 280
   seq_len: 8
   steps_num: 256
   tau: 0.9
   truncate_grads: true
 load_checkpoint: false
-load_path: nn/walker_0_0ep=813.0rew=287.86985430901296
+load_path: path
 model:
   name: continuous_a2c_logstd
 network:
-  cnn:
-    activation: elu
-    convs:
-    - filters: 64
-      kernel_size: 3
-      padding: same
-      strides: 1
-    - filters: 64
-      kernel_size: 3
-      padding: same
-      strides: 1
-    - filters: 64
-      kernel_size: 3
-      padding: same
-      strides: 1
-    - filters: 128
-      kernel_size: 2
-      padding: valid
-      strides: 1
-    initializer:
-      name: variance_scaling_initializer
-      scale: 2
-    regularizer:
-      name: None
-    type: conv1d
   mlp:
-    activation: elu
+    activation: relu
     initializer:
-      name: variance_scaling_initializer
-      scale: 2
+      name: normc_initializer
+      std: 1
     regularizer:
       name: None
     units:
+    - 512
     - 256
     - 128
   name: actor_critic
